Creation of 2 python class templates
created the Classes class and the Group class for the python backbone. The code still needs to be checked to see if there is anything missing or if there are any bugs. I am unable to do so on my side as of right now as I do not have an IDE or text editor that works with Python